+++ lib/market_data/mappers.rb
@@ -3,32 +3,16 @@ require 'market_data/models'
 module MarketData
   module Mappers
     include MarketData::Models
-
     SYMBOL_RESPONSE_KEY = "symbol"
     STATUS_RESPONSE_KEY = "s"
 
     def map_quote response, i=0
-      Quote.new(
-        symbol: response["symbol"][i],
-        ask: response["ask"][i],
-        askSize: response["askSize"][i],
-        bid: response["bid"][i],
-        bidSize: response["bidSize"][i],
-        mid: response["mid"][i],
-        last: response["last"][i],
-        change: response["change"][i],
-        changepct: response["changepct"][i],
-        volume: response["volume"][i],
-        updated: response["updated"][i],
-        high52: response.fetch("high52", nil),
-        low52: response.fetch("low52", nil),
-      )
+      Quote.new(**map_fields_for(response, :quote, i))
     end
 
     def map_bulk_quotes response
       h = Hash.new
-      size = response[SYMBOL_RESPONSE_KEY].size
-      for i in 0..(size - 1) do
+      (0..(response[SYMBOL_RESPONSE_KEY].size - 1)).each do |i|
         qquote = map_quote(response, i)
         h[response[SYMBOL_RESPONSE_KEY][i]] = !qquote.blank? ? qquote : nil
       end
@@ -37,31 +21,50 @@ module MarketData
 
     def map_candles response, symbol
       ar = []
-      size = response["o"].size
-    
-      for i in 0..(size - 1) do
-        ar << Candle.new(
-          open: response["o"][i],
-          high: response["h"][i],
-          low: response["l"][i],
-          close: response["c"][i],
-          volume: response["v"][i],
-          time: response["t"][i],
-          symbol: symbol
-        )
+      (0..(response["o"].size - 1)).each do |i|
+        args = map_fields_for(response, :candle, i)
+        args[:symbol] = symbol
+        ar << Candle.new(**args)
       end
+    
       ar
     end
 
     def map_bulk_candles response
       h = Hash.new
-      size = response[SYMBOL_RESPONSE_KEY].size
-    
-      for i in 0..(size - 1) do
-        candle = Candle.new(symbol: response["symbol"][i], open: response["o"][i], high: response["h"][i], low: response["l"][i], close: response["c"][i], volume: response["v"][i], time: response["t"][i])
+      (0..(response[SYMBOL_RESPONSE_KEY].size - 1)).each do |i|
+        candle = Candle.new(**map_fields_for(response, :candle, i))
         h[response[SYMBOL_RESPONSE_KEY][i]] = !candle.blank? ? candle : nil
       end
       h
+    end
+
+    def map_earning response
+      ar = []
+      (0..(response[SYMBOL_RESPONSE_KEY].size - 1)).each do |i|
+        ar << Earning.new(**map_fields_for(response, :earning, i))
+      end
+      ar
+    end
+
+    def map_fields_for(response, kind, i=0)
+      mapping = {}
+      case kind
+      when :candle
+        mapping = Constants::CANDLE_FIELD_MAPPING
+      when :earning
+        mapping = Constants::EARNING_FIELD_MAPPING
+      when :quote
+        mapping = Constants::QUOTE_FIELD_MAPPING
+      else
+        raise BadParameterError.new("unrecognized model for mapping: #{kind}")
+      end
+      
+      r = {}
+      mapping.each do |field, mapped|
+        r.store(field, response.fetch(mapped, nil).nil? ? nil : response.fetch(mapped)[i])
+      end
+      r
     end
   end
 end